Move cluster_graph
