--- graphrag/index/verbs/graph/clustering/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine graph clustering package root."""
-
-from .cluster_graph import GraphCommunityStrategyType, cluster_graph
-
-__all__ = ["GraphCommunityStrategyType", "cluster_graph"]
