--- graphrag/index/verbs/graph/clustering/strategies/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""Graph Clustering Strategies."""
