--- graphrag/index/verbs/graph/clustering/typing.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""A module containing Communities list definition."""
-
-Communities = list[tuple[int, str, list[str]]]
